Retirando código innecesario

+++ API/src/index.js
@@ -3,33 +3,40 @@ const express = require('express');
 const ngrok = require('ngrok');
 const cors = require('cors');
 const bodyParser = require('body-parser');
-const axios = require('axios');
 const api = express();
 const morgan = require('morgan')
 const fs = require('fs');
-const indexRoutes = require('./routes/index')
 const mongoose = require('mongoose')
 const {mongourl} = require('./config/keys')
-const Testy = require('./models/testy')
 const SnapModel = require('./models/snap');
-const { db } = require('./models/testy');
-var jsonParser = bodyParser.json();
 var urlencodedParser = bodyParser.urlencoded({ extended: false});
-var ObjectId = require('mongodb').ObjectId;
-var o_id;
-/*var imageJSON = JSON.parse(fs.readFileSync('C:/Users/sergi/Downloads/my_outfile.json', 'utf8'))
-var imageString = imageJSON['image'].substring(23, imageJSON['image'].length)
+
 
 try{
-    fs.writeFile('C:/Users/sergi/OneDrive/Escritorio/White Box 2/Tec2/Last Chance/Inteligencia Artificial/Proyecto/SnapAI/Inteligencia/image_to_read/image.jpg', imageString, 'base64', function(err){
-        if (err) throw err
-        console.log('File saved.')
-    })
-    fs.unlinkSync('C:/Users/sergi/Downloads/my_outfile.json')
-    console.log("File removed, check it out")
+    if(fs.existsSync('C:/Users/sergi/Downloads/my_outfile.json')){
+        var imageJSON = JSON.parse(fs.readFileSync('C:/Users/sergi/Downloads/my_outfile.json', 'utf8'))
+        var imageString = imageJSON['image'].substring(23, imageJSON['image'].length)
+        fs.writeFile('C:/Users/sergi/OneDrive/Escritorio/White Box 2/Tec2/Last Chance/Inteligencia Artificial/Proyecto/SnapAI/Inteligencia/image_to_read/image.jpg', imageString, 'base64', function(err){
+            if (err) throw err
+            console.log('File saved.')
+
+        })
+        fs.unlinkSync('C:/Users/sergi/Downloads/my_outfile.json')
+        console.log("File removed, check it out")
+    }else{
+        console.log("No existe el archivo")
+    }
+
+    if(fs.existsSync('C:/Users/sergi/Downloads/new_entry.json')){
+        var new_entry = JSON.parse(fs.readFileSync('C:/Users/sergi/Downloads/new_entry.json', 'utf8'))
+    }else{
+        console.log("No existe el archivo new_entry")
+    }
 }catch(err){
-    console.error(err)
-}*/
+    console.log("El archivo no existe")
+    console.log(err)
+}
+
 mongoose.connect(mongourl);
 
 api.listen(8080, () => {
@@ -81,20 +88,22 @@ api.get('/nobody', (req, res) => {
 
 
 
-api.get('/save', (req, res) =>{
-    const Item = new SnapModel(imageJSON)
-    Item.save().then(data=>{
-        console.log("Registrado compa")
-    }).catch(err=>{
-        throw err;
-    })
-})
 
 
 const updateModel = require('./models/updateModel');
 const userModel = require('./models/userModel');
 var toSendUpdate;
 var toSendUser;
+
+api.get('/save', (req, res) =>{
+    const Item = new SnapModel(new_entry)
+    Item.save().then(data=>{
+        console.log("Registrado")
+    }).catch(err=>{
+        throw err;
+    })
+})
+
 userModel.find({}, (err, oneModel) => {
     if(err){
         console.log(err);
@@ -145,9 +154,6 @@ api.get('/creando', urlencodedParser,async function(req, res){
     })
 });
 */
-(async () => {
-    console.log("Async async")
-})
 
 api.post('/update', urlencodedParser,function(req, res){
     if(!req.body) return res.sendStatus(400)
@@ -177,33 +183,29 @@ api.post('/update', urlencodedParser,function(req, res){
             .filter(dirent => dirent.isDirectory())
             .map(dirent => dirent.name)
 
-        const copyDirectory = getDirectories("C:/Users/sergi/OneDrive/Escritorio/White Box 2/Tec2/Last Chance/Inteligencia Artificial/Proyecto/SnapAI/Inteligencia/labeled_images");
-        for(var i in copyDirectory) console.log(copyDirectory[i]);
-        //Create the json file to update the name array (àrt of the knowledge base)
+        const auxDirectory = getDirectories("C:/Users/sergi/OneDrive/Escritorio/White Box 2/Tec2/Last Chance/Inteligencia Artificial/Proyecto/SnapAI/Inteligencia/labeled_images");
+        var copyDirectory = `[ "${auxDirectory[0]}" `;
 
+        for(var i = 1 in auxDirectory){
+            copyDirectory += `,  "${auxDirectory[i]}"`
+        }
+        copyDirectory += `]`
+
+        for(var i = 0 in req.body){
+            console.log(req.body.id)
+        }
+        //Guardar lista de nombres
+        fs.writeFile(`C:/Users/sergi/OneDrive/Escritorio/White Box 2/Tec2/Last Chance/Inteligencia Artificial/Proyecto/SnapAI/Inteligencia/new_json_data/new_data.json`, copyDirectory, 'UTF-8', function(err){
+            if (err) throw err
+            console.log('New set of names saved.')
+        })
     }catch(err){
         console.error(err)
     }
+    //Guardar id al que pertenece
+    fs.writeFile(`C:/Users/sergi/OneDrive/Escritorio/White Box 2/Tec2/Last Chance/Inteligencia Artificial/Proyecto/SnapAI/Inteligencia/new_json_data/id.json`, `${req.body.id}`, 'UTF-8', function(err){
+        if (err) throw err
+        console.log('Id saved')
+    })
+
 });
-
-//Get the directory list
-const {readdirSync} = require('fs')
-const getDirectories = source =>
-readdirSync(source, { withFileTypes: true })
-    .filter(dirent => dirent.isDirectory())
-    .map(dirent => dirent.name)
-
-const auxDirectory = getDirectories("C:/Users/sergi/OneDrive/Escritorio/White Box 2/Tec2/Last Chance/Inteligencia Artificial/Proyecto/SnapAI/Inteligencia/labeled_images");
-var copyDirectory = `new_set_names = "[{ 'name': '${auxDirectory[0]}' }`;
-
-for(var i = 1 in auxDirectory){
-    copyDirectory += `, { 'name' : '${auxDirectory[i]}'}`
-}
-copyDirectory += `]"`
-
-console.log(copyDirectory);
-
-fs.writeFile(`C:/Users/sergi/OneDrive/Escritorio/White Box 2/Tec2/Last Chance/Inteligencia Artificial/Proyecto/SnapAI/Inteligencia/new_json_data/new_data.json`, copyDirectory, 'UTF-8', function(err){
-    if (err) throw err
-    console.log('New set of names saved.')
-})
